services/common/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "kg_common"
version = "0.1.0"
dependencies = [
  "rdflib==7.0.0",
  "requests>=2.32.2",
  "fastapi==0.111.0",
  "uvicorn==0.30.1",
  "pydantic==2.7.1",
  "qdrant-client==1.9.3",
  "sentence-transformers==3.0.1",
  "spacy==3.7.5",
  "rapidfuzz==3.9.3",
  "llama-cpp-python==0.2.84",
  "celery==5.3.6",
  "redis==5.0.7",
  "pdfminer.six==20231228",
  "python-multipart==0.0.9",
  "tqdm==4.66.4",
  "orjson==3.10.3",
  "prometheus-client==0.20.0",
]
